symbols: Add tests for key merge modes (keysyms/actions)
This commit adds tests for merging various key configurations:
- With/without keysyms/actions
- Single/multiple keysyms/actions per level

We test all the merge modes for including a map:
- default (include)
- augment
- override
- replace

We also test against xkbcomp so we can compare its handling of merge
modes with xkbcommon. Unfortunately, there are currently some
differences due to xkbcomp discarding trailing `NoSymbol`. Thus the
tests are split into: common, X11-only and xkbcommon-only.

The tests are generated with:
- A Python scripts `scripts/update-symbols-tests.py` for symbols data
  and C headers (key sequence tests). Note that these files are used
  only to validate the following files.
- The `merge-modes` test generates its own files for golden test using:
  `build/test-merge-modes update`.